Remove the decimal integer patch
This bug was patched years ago by some helpful developer

Merge input validation onto the Command object

It was done in a patch for legacy reasons. However as this is now
a hard fork, it is easier to maintain on to Command object itself

Merge priority sort onto the commander
