--- lib/commander/patches/decimal-integer.rb
@@ -1,17 +1,0 @@
-# Patches the underling OptionParser DecimalInteger type so that it doesn't
-# convert numbers starting with a '0' as if they are an Octal number
-module Commander
-  module Patches
-    module DecimalInteger
-      decimal = '\d+(?:_\d+)*'
-      DecimalInteger = /\A[-+]?#{decimal}\z/io
-      ::OptionParser::accept(DecimalInteger, DecimalInteger) {|s,|
-        begin
-          Integer(s, 10)
-        rescue ArgumentError
-          raise ::OptionParser::InvalidArgument, s
-        end if s
-      }
-    end
-  end
-end
